+++ C-edition/src/functions.c
@@ -13,12 +13,15 @@ void help() {
     echo("\e[1mAVAILABLE OPTIONS:\e[0m");
     echo("	install - install a package.");
     echo("	remove - uninstall a package.");
+	echo("	purge - uninstall a package along with its configuration files.");
     echo("	search - search a package.");
     echo("	find - (pacman -F) in debian: 'apt-file search'.");
     echo("	update - equivalent of 'apt update' in debian.");
 	echo("	upgrade - equivalent of 'apt upgrade' in debian.");
     echo("	autoclean - clean up all local caches.");
     echo("	autoremove - remove packages that are no longer needed.");
+    echo("	show - show the information of a package that is installed.");
+	echo("	show-all - same as 'show', but shows packages from the repos.");
 	echo("	list-installed - list all installed packages.");
     echo("	help - show this help.");
 	echo("	version - show version and about information.\n");