+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - Versions 0.1-alpha and 0.1.1-alpha released.
 - Theoretically the script is fully functional
 ### 03/5/2021
-- Split the `update` command to `update` and `upgrade` so its more similar to PT.
+- Split the `update` command to `update` and `upgrade` so its more similar to apt.
 - Version 0.3.1-alpha
 <hr>
+### 09/5/2021
+- make C edition as good as the bash edition.
+- add 'purge' option to bash and C edition.
+- fix typo in the help function in the C and bash editions.
+- add 'show' and 'show-all' commands (C and bash editions).
+